[chg] redis rpc cache use unlink
- https://redis.io/commands/unlink/
- https://redis.com/blog/the-little-known-feature-of-redis-4-0-that-will-speed-up-your-applications/

> As you might have guessed, it’s UNLINK to the rescue! UNLINK is
syntactically the same as DEL but provides a much more ideal solution.
First it removes the key from the overall keyspace. Then, in a different
thread it starts reclaiming the memory. This is a safe operation from a
multithreaded perspective since it (in the main thread) removes the item
from the keyspace and thus makes it inaccessible from any Redis command.